+++ scripts/deployments.json
@@ -45,7 +45,6 @@
 		"BinaryOptionsMarketFactory": "0x0314824253cEc50d144d60CfEA3742874Df72B92",
 		"BinaryOptionMarketData": "0x42Af0890d254387110555A46d261B9e4f6E327Ac",
 		"OpThales_L1": "0x008e86b2e21ce8d71C4700A9a174D151839cFbc7",
-		"ThalesExchanger": "0x7b9a3a5599B26a3243b50cf6e1506f7bE180aC3c",
 		"OwnedUpgradeabilityProxy": "0x7596C9aA9ce105610324AcC27D943C5235518701",
 		"ProxyThalesExchanger": "0x7596C9aA9ce105610324AcC27D943C5235518701"
 	},
