+++ scripts/l1_l2_exchanger/deprecated/l1_l2_exchange.js
@@ -16,7 +16,7 @@ const {
 	currentTime,
 	multiplyDecimalRound,
 	divideDecimalRound,
-} = require('../../test/utils/index')();
+} = require('../../../test/utils/index')();
 
 const ZERO_ADDRESS = '0x' + '0'.repeat(40);
 
@@ -25,14 +25,14 @@ const L1_MESSENGER_ADDRESS = '0x4361d0F75A0186C05f971c566dC6bEa5957483fD';
 const STATE_COMMITMENT_CHAIN_ADDRESS = '0xD7754711773489F31A0602635f3F167826ce53C5';
 const old_tx_hash = '0xd84ecfc22b3ce5b4605abcd5170e69da701e8d189fa10a52fa82533d4e59e527'
 
-const { getTargetAddress, setTargetAddress } = require('../helpers');
+const { getTargetAddress, setTargetAddress } = require('../../helpers');
 
 const L2StandardBridgeArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L2StandardBridge.sol/OVM_L2StandardBridge.json');
 const L1StandardBridgeArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/tokens/OVM_L1StandardBridge.sol/OVM_L1StandardBridge.json');
 const L1CrossDomainMessengerArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/messaging/OVM_L1CrossDomainMessenger.sol/OVM_L1CrossDomainMessenger.json');
 const L2CrossDomainMessengerArtifacts = require('@eth-optimism/contracts/artifacts/contracts/optimistic-ethereum/OVM/bridge/messaging/OVM_L2CrossDomainMessenger.sol/OVM_L2CrossDomainMessenger.json');
 
-const { toBytes32 } = require('../../index');
+const { toBytes32 } = require('../../../index');
 
 async function main() {
 	let accounts = await ethers.getSigners();
@@ -62,7 +62,7 @@ async function main() {
 	console.log("Optimistic Kovan block number: ", blockNumber);
 	
 	const ThalesAddress = getTargetAddress('Thales', net_kovan);
-	const ThalesExchangerAddress = getTargetAddress('ThalesExchanger', net_kovan);
+	const ThalesExchangerAddress = getTargetAddress('ProxyThalesExchanger', net_kovan);
 	const OP_Thales_L1Address = getTargetAddress('OpThales_L1', net_kovan);
 	const OP_Thales_L2Address = getTargetAddress('OpThales_L2', net_optimistic_kovan);
 
@@ -70,7 +70,7 @@ async function main() {
 	const L2StandardBridge = new ethers.ContractFactory(L2StandardBridgeArtifacts.abi, L2StandardBridgeArtifacts.bytecode);
 	const L1StandardBridge = new ethers.ContractFactory(L1StandardBridgeArtifacts.abi, L1StandardBridgeArtifacts.bytecode);
 	const OP_Thales_L1 = await ethers.getContractFactory('/contracts/Token/OpThales_L1.sol:OpThales');
-	const ThalesExchanger = await ethers.getContractFactory('ThalesExchanger');
+	const ThalesExchanger = await ethers.getContractFactory('ProxyThalesExchanger');
 	const OP_Thales_L2 = await ethers.getContractFactory('/contracts/Token/OpThales_L2.sol:OpThales');
 	const Thales = await ethers.getContractFactory('Thales');
