+++ scripts/l1_l2_exchanger/deprecated/local_deploy_l1_l2_exchange.js
