+++ scripts/l1_l2_exchanger/deprecated/verify_L1_L2.js
